+++ outputs.tf
@@ -22,8 +22,3 @@ output "subnet_names" {
   description = "Names list of the created subnet"
   value       = azurerm_subnet.subnet[*].name
 }
-
-output "subnet_ip_configurations" {
-  description = "The collection of IP Configurations with IPs within this subnet"
-  value       = azurerm_subnet.subnet[*].ip_configurations
-}